server/pyproject.toml
[tool.poetry]
name = "wildberriestracker"
version = "0.1.0"
description = ""
authors = ["AlekseyAmp <thientaovv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.27.1"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.27"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
pytz = "^2024.1"
autopep8 = "^2.0.4"
flake8 = "^7.0.0"
isort = "^5.13.2"
pydantic = "2.0.2"
fastapi = "^0.109.2"
celery = "^5.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
